The problem statement
Create a vector called x and place values 1 to 100. Create a y vector that takes the square root of the numbers in x. Plot x and y as shown below. Your task is to create the mathematical expressions displayed on the graph. Hint: the function to use is expression(). Once you get the expression, use text(x,y,label) to place the formulas.
 
  
                       
 
Save all the code above in a script file called A5yourname. Make sure to include ample comments to describe your code. Upload your script to submit the assignment.