Problem Statement
Create the following functions. You don’t have to save the functions to a separate file. After creating the functions, test them with at least two sample inputs to the functions.

Create a function that return the length of the hypotenuse of a right angle triangle when the length of the two sides are given as the parameters of the function.
Create a function that takes in a numeric vector. The function should do the following: (1) takes all values below the mean and calculates the median of these, (2) takes all values above the mean and calculates the median of these, and (3) return the difference in the value between the higher and the lower median.
Create a function that takes in a numeric vector. The output should be a vector with running median values. The ith element of the output vector should be the median of the values in the input vector from 1 to i.
There is a package with a function that enables you to check if a number is prime.
install.packages("schoolmath")

library(schoolmath)

is.prim(3)

Create a function that takes in two integers (set default values of 1 to both). The function should calculate the number of prime numbers between the two values.

Simulate a function to roll a dice. Note that a dice turns up with numbers 1, 2, 3, 4, 5 or 6. The function should do the following: you roll the dice twice, add up the two numbers that come up, and return the sum. 
Hint: use the function sample(vector, size, replace=T/F). The argument vector indicates the vector of values to draw from. The argument size indicates how many values to draw, and replace argument indicates whether you sample with replacement. Also note that this function does not take in any parameters, it returns a string, i.e. if the result of two rolling are 3 and 6, the function should return "The two values are 3 and 6, so the sum of them is 9". Hint: Use paste() function